Added leave donut system.

+++ one-more-time/src/resources.rs
@@ -2,6 +2,6 @@ use bevy::prelude::*;
 
 #[derive(Resource)]
 pub struct GameState {
-    pub balance: i32,
+    pub balance: f32,
     pub cook_donut_count: u8,
 }
